add archive mode

+++ kopano-cleanup/kopano-cleanup.py
@@ -11,10 +11,11 @@ from MAPI.Util import *
 
 
 def opt_args():
-    parser = kopano.parser('skpcfm')
+    parser = kopano.parser('skpcfmUP')
     parser.add_option("--user", dest="user", action="store", help="Run script for user")
     parser.add_option("--wastebasket", dest="wastebasket", action="store_true",
                       help="Run cleanup script for the wastebasket folder")
+    parser.add_option("--archive", dest="archive", action="store", help="instead of removing items archive them into this folder")
     parser.add_option("--junk", dest="junk", action="store_true", help="Run cleanup script for the junk folder")
     parser.add_option("--force", dest="force", action="store_true", help="Force items without date to be removed")
     parser.add_option("--days", dest="days", action="store", help="Delete older then x days")
@@ -28,8 +29,8 @@ def progressbar(folder, daysbeforedeleted):
     try:
         from progressbar import Bar, AdaptiveETA, Percentage, ProgressBar
     except ImportError:
-        print '''Please download the progressbar library from https://github.com/niltonvolpato/python-progressbar or
-        run without the --progressbar parameter'''
+        print('''Please download the progressbar library from https://github.com/niltonvolpato/python-progressbar or
+        run without the --progressbar parameter''')
         sys.exit(1)
     widgets = [Percentage(),
                ' ', Bar(),
@@ -44,8 +45,16 @@ def progressbar(folder, daysbeforedeleted):
     return pbar
 
 
-def deleteitems(options, folder, daysbeforedeleted, pbar):
+def deleteitems(options, user, folder):
     itemcount = 0
+    daysbeforedeleted = datetime.datetime.now() - timedelta(days=int(options.days))
+    pbar = None
+    if options.progressbar:
+        pbar = progressbar(folder, daysbeforedeleted)
+
+    if options.archive:
+        archive_folder = user.store.folder(options.archive, create=True)
+
     for item in folder.items():
         date = None
         if not item.prop(PR_LAST_MODIFICATION_TIME).value and options.force:
@@ -55,44 +64,58 @@ def deleteitems(options, folder, daysbeforedeleted, pbar):
         if date:
             if date <= daysbeforedeleted:
                 if options.verbose:
-                    print 'Deleting \'%s\'' % item.subject
+                    if options.archive:
+                        print('Archiving \'{}\''.format(item.subject))
+                    else:
+                        print('Deleting \'{}\''.format(item.subject))
+
                 if not options.dryrun:
-                    folder.delete(item)
+                    if options.archive:
+                        folder.move(item, archive_folder)
+                    else:
+                        folder.delete(item)
                 if pbar:
                     pbar.update(itemcount + 1)
                 itemcount += 1
+
+    if options.progressbar:
+        pbar.finish()
+    if options.archive:
+        print('Archived {} item(s) for user \'{}\' in folder \'{}\' to folder \'{}\''.format(itemcount, user.name,
+                                                                                             folder.name,
+                                                                                             archive_folder.name))
+    else:
+        print('Deleted {}  item(s) for user \'{}\' in folder \'{}\''.format(itemcount, user.name, folder.name))
+
     return itemcount
+
 
 
 def main():
     options, args = opt_args()
     if not options.user or not options.days:
-        print 'Please use:\n %s --user <username> --days <days> ' % (sys.argv[0])
+        print('Please use:\n {} --user <username> --days <days> ',format(sys.argv[0]))
         sys.exit(1)
 
-    daysbeforedeleted = datetime.datetime.now() - timedelta(days=int(options.days))
+
 
     user = kopano.Server(options).user(options.user)
-    print 'Running script for \'%s\'' % user.name
-    pbar = None
+    print('Running script for \'{}\''.format(user.name))
+
     if options.wastebasket:
-        folder = user.store.wastebasket
-        if options.progressbar:
-            pbar = progressbar(folder, daysbeforedeleted)
-        count = deleteitems(options, folder, daysbeforedeleted, pbar)
-        if options.progressbar:
-            pbar.finish()
-        print 'Deleted %s item(s) for user \'%s\' in folder \'%s\'' % (count, user.name, folder.name)
+            folder = user.store.wastebasket
+            deleteitems(options, user, folder)
 
     if options.junk:
-        folder = user.store.junk
-        if options.progressbar:
-            pbar = progressbar(folder, daysbeforedeleted)
-        count = deleteitems(options, folder, daysbeforedeleted, pbar)
-        if options.progressbar:
-            pbar.finish()
-        print 'Deleted %s item(s) for user \'%s\' in folder \'%s\'' % (count, user.name, folder.name)
+            folder = user.store.junk
+            deleteitems(options, user, folder)
 
+    '''
+    Loop over all the folders that are passed with the -f parameter 
+    '''
+    if options.folders:
+        for folder in user.store.folders(options.folders):
+            deleteitems(options, user, folder)
 
 if __name__ == "__main__":
     main()
